Move InvalidDimensionsException to net.imglib2.exception package
